@@ -43,33 +45,29 @@
   "deprecated": false,
   "description": "RFC4122 (v1, v4, and v5) UUIDs",
   "devDependencies": {
-    "@babel/cli": "7.11.6",
-    "@babel/core": "7.11.6",
-    "@babel/preset-env": "7.11.5",
-    "@commitlint/cli": "11.0.0",
-    "@commitlint/config-conventional": "11.0.0",
-    "@rollup/plugin-node-resolve": "9.0.0",
-    "babel-eslint": "10.1.0",
-    "bundlewatch": "0.3.1",
-    "eslint": "7.10.0",
-    "eslint-config-prettier": "6.12.0",
-    "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.22.1",
+    "@babel/cli": "7.18.10",
+    "@babel/core": "7.18.10",
+    "@babel/eslint-parser": "7.18.9",
+    "@babel/preset-env": "7.18.10",
+    "@commitlint/cli": "17.0.3",
+    "@commitlint/config-conventional": "17.0.3",
+    "bundlewatch": "0.3.3",
+    "eslint": "8.21.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-config-standard": "17.0.0",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "3.1.4",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
-    "husky": "4.3.0",
-    "jest": "25.5.4",
-    "lint-staged": "10.4.0",
+    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-promise": "6.0.0",
+    "husky": "8.0.1",
+    "jest": "28.1.3",
+    "lint-staged": "13.0.3",
     "npm-run-all": "4.1.5",
     "optional-dev-dependency": "2.0.1",
-    "prettier": "2.1.2",
+    "prettier": "2.7.1",
     "random-seed": "0.3.0",
-    "rollup": "2.28.2",
-    "rollup-plugin-terser": "7.0.2",
-    "runmd": "1.3.2",
-    "standard-version": "9.0.0"
+    "runmd": "1.3.6",
+    "standard-version": "9.5.0"
   },
   "exports": {
     ".": {
