@@ -9,6 +9,7 @@ const bodyParser = require('body-parser')
 
 
 const UserController = require('./controllers/user');
+const DataController = require('./controllers/data');
 
 // Load environment variables from the .env file
 dotenv.config();
@@ -17,17 +18,27 @@ dotenv.config();
 const app = express();
 app.use(bodyParser.json())
 
+//user control routes
+app.post('/register', UserController.Register); //register route 
+app.post('/login',UserController.Login); //login route
 
-app.post('/register', UserController.create);
+//data control routes
+app.post('/create',DataController.createData); //create route
+app.put('/edit',DataController.editData); //edit route
+app.get('/view', DataController.ViewData); //view all
+app.post('/viewone', DataController.ViewOneData); //view special
+app.delete('/delete/:id', DataController.Delete); //delete data
+
 // app.post('/login', (req,res) => {
 //     const {email,password} = req.body;
 //     console.log(email,password)
 // })
 
+//beacause of an issue when reading process.env, used mongo url directly in index.ts
 const port = process.env.PORT || 3000;
 const mongoUrl = "mongodb+srv://user123:123123123@cluster0.xkikzml.mongodb.net/?retryWrites=true&w=majority"
 mongoose.connect(mongoUrl, { bufferCommands: false });
-// mongoose.set('strictQuery', false);
+mongoose.set('strictQuery', false);
 // mongoose.set('strictQuery', true);
 
 // Connect to the MongoDB instance
@@ -37,7 +48,6 @@ mongodb.MongoClient.connect(mongoUrl, (error, client) => {
     return;
   }
   console.log("MongoDB connected")
-  // Set up routes and middleware here
 
   // Start the server
   app.listen(port, () => {
